move the wait for workers login into http_server
